+++ src/test/java/lotto/LottoTest.java
@@ -1,11 +1,11 @@
 package lotto;
 
+import lotto.Generator.NumberGenerator;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
-import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
